+++ Pandora/Paradigm/Controlflow.hs
@@ -2,4 +2,4 @@ module Pandora.Paradigm.Controlflow (module Exports) where
 
 import Pandora.Paradigm.Controlflow.Pipeline as Exports
 import Pandora.Paradigm.Controlflow.Observable as Exports
-import Pandora.Paradigm.Controlflow.Joint as Exports
+import Pandora.Paradigm.Controlflow.Effect as Exports