+++ Pandora/Paradigm/Schemes/TUVW.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Schemes.TUVW (TUVW (..)) where
 
 import Pandora.Core.Functor (type (:.), type (:=))
-import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, run))
+import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run))
 
 newtype TUVW ct cu cv cw t u v w a = TUVW (t :. u :. v :. w := a)